[MSG] Added ability to query for message snippets of a conversation
